Move store API to separate file
