+++ platforms/ESP-IDF/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(SOURCE_FILES
         ../../src/Debug/debugger.cpp
         ../../src/Interpreter/instructions.cpp
+        ../../src/Interpreter/wasm.cpp
         ../../src/Memory/mem.cpp
         ../../src/Primitives/idf.cpp
         ../../src/Edward/proxy.cpp
