+++ src/Interpreter/wasm.h
@@ -1,0 +1,3 @@
+void report_overflow(Module *m, uint8_t *maddr);
+
+bool store(Module *m, uint8_t type, uint32_t addr, StackValue &sval);
